Limit tGATT_IF range
This commit restricts the range of tGATT_IF
values to 1-248, reserving 249-255 for special use cases.
This change ensures consistency and avoids potential conflicts.

Additionally, the commit addresses issues found while
checking all the tGATT_IF usages.

Bug: 348559823
Bug: 273561907
Test: atest BumbleBluetoothTests

+++ system/bta/dm/bta_dm_gatt_client.cc
@@ -54,7 +54,7 @@ gatt_interface_t default_gatt_interface = {
                 [](tGATT_IF client_if, const RawAddress& remote_bda, bool is_direct) {
                   gatt_history_.Push(base::StringPrintf(
                           "%-32s bd_addr:%s client_if:%hu is_direct:%c", "GATTC_CancelOpen",
-                          ADDRESS_TO_LOGGABLE_CSTR(remote_bda), client_if,
+                          ADDRESS_TO_LOGGABLE_CSTR(remote_bda), static_cast<uint16_t>(client_if),
                           (is_direct) ? 'T' : 'F'));
                   BTA_GATTC_CancelOpen(client_if, remote_bda, is_direct);
                 },
@@ -101,8 +101,8 @@ gatt_interface_t default_gatt_interface = {
                    tBTM_BLE_CONN_TYPE connection_type, bool opportunistic, uint16_t preferred_mtu) {
                   gatt_history_.Push(base::StringPrintf(
                           "%-32s bd_addr:%s client_if:%hu type:0x%x opportunistic:%c", "GATTC_Open",
-                          ADDRESS_TO_LOGGABLE_CSTR(remote_bda), client_if, connection_type,
-                          (opportunistic) ? 'T' : 'F'));
+                          ADDRESS_TO_LOGGABLE_CSTR(remote_bda), static_cast<uint16_t>(client_if),
+                          connection_type, (opportunistic) ? 'T' : 'F'));
                   BTA_GATTC_Open(client_if, remote_bda, BLE_ADDR_PUBLIC, connection_type,
                                  BT_TRANSPORT_LE, opportunistic, LE_PHY_1M, preferred_mtu);
                 },
